Added code log msg_id from user

+++ bot/bot_func.py
@@ -11,11 +11,17 @@ from meta.user import TodaysTasks, UserQueueData, UserCommands, UserCmd, User
 from meta.handlers import UserCommandHandler, UserTaskHandler
 
 from database.database import Database
-
+from database.message_id import MessageID
 
 bf_log = logging.getLogger("bot_func")
 bf_logger = Logger(bf_log, base_level=logging.DEBUG, filename="")
 db = Database()
+msg_id_db = MessageID()
+
+
+def send_msg(bot: telebot.TeleBot, user_id: int, message: str) -> None:
+    rv = bot.send_message(user_id, text=message)
+    msg_id_db.add_msg_id(user_id, rv.message_id)
 
 
 def parse_text(
@@ -39,7 +45,7 @@ def parse_text(
                     cmd=UserCommands.TASKDESC,
                 ),
             )
-            bot.send_message(_id, text="Enter the task description")
+            send_msg(bot, _id, message="Enter the task description")
 
         elif text == "removetasks":
             cmd_handler.set_user_command(
@@ -49,7 +55,7 @@ def parse_text(
                     cmd=UserCommands.REMOVETASK,
                 ),
             )
-            bot.send_message(_id, text="Enter the Task ID of the task to remove..")
+            send_msg(bot, _id, message="Enter the Task ID of the task to remove..")
 
         elif text == "viewtasks":
             cmd_handler.set_user_command(
@@ -64,13 +70,13 @@ def parse_text(
             if tasks:
                 tasks = tasks.tasks
                 for task in tasks:
-                    bot.send_message(_id, text=str(task))
+                    send_msg(bot, _id, message=str(task))
 
         else:
             if cmd_handler.get_user_command(_id):
                 if cmd_handler.get_user_command(_id).cmd == UserCommands.TASKDESC:
                     task_handler.add_task(_id, task_desc=data.text)
-                    bot.send_message(_id, text="Enter the date")
+                    send_msg(bot, _id, message="Enter the date")
                     cmd_handler.set_user_command(
                         _id,
                         cmd=UserCmd(
@@ -91,25 +97,27 @@ def parse_text(
                             username=user_q_data.username, user_id=_id, tasks=[task]
                         )
                         db.add_task(user)
-                        bot.send_message(_id, text="Task added successfully")
+                        send_msg(bot, _id, message="Task added successfully")
                         cmd_handler.set_user_command(
                             _id,
                             UserCmd(time_inserted=time_e, cmd=UserCommands.TASKDONE),
                         )
                     except ValueError:
-                        bot.send_message(_id, f"Date should be of the form DD-MM-YYYY")
+                        send_msg(
+                            bot, _id, message="Date should be of the form DD-MM-YYYY"
+                        )
 
                 if cmd_handler.get_user_command(_id).cmd == UserCommands.REMOVETASK:
                     try:
                         db.delete_user_tasks(_id, task_id=int(data.text))
-                        bot.send_message(_id, text="Task removed successfully..")
+                        send_msg(bot, _id, message="Task removed successfully..")
                     except ValueError:
-                        bot.send_message(_id, text="Task ID must be a number..")
+                        send_msg(bot, _id, message="Task ID must be a number..")
 
 
 def send_cmd_time_out(bot: telebot.TeleBot, users: List[int]):
     for user in users:
-        bot.send_message(user, text="Ooops. Your last command/ task timed out")
+        send_msg(bot, user, message="Ooops. Your last command or task timed out")
 
 
 def todays_tasks_func(bot: telebot.TeleBot, bot_logger) -> List[TodaysTasks]:
@@ -117,7 +125,10 @@ def todays_tasks_func(bot: telebot.TeleBot, bot_logger) -> List[TodaysTasks]:
     bot_logger.log(logging.INFO, message=f"Sending todays tasks {tasks=}")
 
     for task in tasks:
-        bot.send_message(task.user_id, text="Today's Task's")
-        for t in task.tasks:
-            bot.send_message(task.user_id, str(t))
-            db.delete_user_tasks(task.user_id, t.task_id)
+        send_msg(bot, task.user_id, message="Today's Task's")
+        if not task.tasks:
+            send_msg(bot, task.user_id, message="You don't have any tasks..")
+        else:
+            for t in task.tasks:
+                send_msg(bot, task.user_id, message=str(t))
+                db.delete_user_tasks(task.user_id, t.task_id)
